+++ src/main/java/com/bl/dao/GeneralDAO.java
@@ -10,8 +10,8 @@ public interface GeneralDAO {
 	public List<GeneralDTO> colorsList() ;
 	public GeneralDTO colorNameById(Integer id) ;
 	public List<GeneralDTO> countryList() ;
-	public List<GeneralDTO> governorateList(Long countryId) ;
-	public List<GeneralDTO> cityDistrictList(Long governorateId) ;
-	public List<CountryGovernorateCityDistrictDTO> findGovernorateIdAndCountryIdByCityDistrictId(Long cityDistrictId) ;
+	public List<GeneralDTO> governorateList(Integer countryId) ;
+	public List<GeneralDTO> cityDistrictList(Integer governorateId) ;
+	public List<CountryGovernorateCityDistrictDTO> findGovernorateIdAndCountryIdByCityDistrictId(Integer cityDistrictId) ;
 	public GeneralDTO currencyById(Integer id) ;
 }
